Cleaned up in some interfaces in app frontend
