+++ .vscode/settings.json
@@ -28,5 +28,8 @@
   "eslint.workingDirectories": [
     { "directory": "./src/studio/src/designer/frontend", "changeProcessCWD": true },
     { "directory": "./src/Altinn.Apps/AppFrontend/react", "changeProcessCWD": true}
-  ]
+  ],
+  "editor.codeActionsOnSave": {
+    "source.fixAll": true,
+  },
 }
