+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/__tests__/utils/rules.test.tsx
@@ -3,7 +3,7 @@
 import 'jest';
 import { IFormData } from '../../src/features/form/data/formDataReducer';
 import { ILayoutComponent } from '../../src/features/form/layout';
-import { IDataModelFieldElement } from '../../src/features/form/rules';
+import { IDataModelFieldElement } from '../../src/types';
 import { checkIfRuleShouldRun, getRuleModelFields } from '../../src/utils/rules';
 
 describe('>>> features/rules checkIfRuleShouldRun', () => {
