+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "altinn-app-frontend",
-  "version": "2.1.11",
+  "version": "2.1.12",
   "description": "",
   "main": "index.js",
   "scripts": {
