+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/features/form/dynamics/conditionalRendering/conditionalRenderingSagas.ts
@@ -1,7 +1,6 @@
 import { SagaIterator } from 'redux-saga';
 import { all, call, select, take, takeLatest } from 'redux-saga/effects';
-import { IRuntimeState } from '../../../../types';
-import { IValidations, IUiConfig } from '../../../../types/global';
+import { IRuntimeState, IValidations, IUiConfig } from 'src/types';
 import { runConditionalRenderingRules } from '../../../../utils/conditionalRendering';
 import * as FormDataActionTypes from '../../data/formDataActionTypes';
 import { IFormData } from '../../data/formDataReducer';
@@ -28,8 +27,8 @@ function* checkIfConditionalRulesShouldRunSaga(): SagaIterator {
     const formValidations: IValidations = yield select(FormValidationSelector);
     const uiConfig: IUiConfig = yield select(UiConfigSelector);
     const componentsToHide: string[] = runConditionalRenderingRules(
-        conditionalRenderingState,
-        formData,
+      conditionalRenderingState,
+      formData,
     );
 
     if (shouldHidddenFieldsUpdate(uiConfig.hiddenFields, componentsToHide)) {
@@ -56,13 +55,12 @@ export function* waitForAppSetupBeforeRunningConditionalRulesSaga(): SagaIterato
     take(FormLayoutActionTypes.FETCH_FORM_LAYOUT_FULFILLED),
     take(FormDataActionTypes.FETCH_FORM_DATA_FULFILLED),
     take(FormDynamicsActionTypes.FETCH_SERVICE_CONFIG_FULFILLED),
-    take(RulesActionTypes.FETCH_RULE_MODEL_FULFILLED)
+    take(RulesActionTypes.FETCH_RULE_MODEL_FULFILLED),
   ]);
   yield call(checkIfConditionalRulesShouldRunSaga);
 }
 
 function shouldHidddenFieldsUpdate(currentList: string[], newList: string[]): boolean {
-  
   if (!currentList || currentList.length !== newList.length) {
     return true;
   }
@@ -71,7 +69,7 @@ function shouldHidddenFieldsUpdate(currentList: string[], newList: string[]): bo
     return true;
   }
 
-  if (currentList.find(element => !newList.includes(element))) {
+  if (currentList.find((element) => !newList.includes(element))) {
     return true;
   }
