+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/features/form/rules/index.d.ts
@@ -1,27 +1,3 @@
-export interface IDataModelFieldElement {
-  choices?: any;
-  customProperties?: any;
-  dataBindingName: string;
-  displayString: string;
-  fixedValue?: any;
-  id: string;
-  isReadOnly: boolean;
-  isTagContent: boolean;
-  jsonSchemaPointer: string;
-  maxOccurs: number;
-  minOccurs: number;
-  name: string;
-  parentElement: string;
-  restrictions: any;
-  texts: any;
-  type: string;
-  typeName?: string;
-  xmlSchemaXPath: string;
-  xName?: string;
-  xPath: string;
-  xsdValueType?: string;
-}
-
 export interface IRuleModelFieldElement {
   type: string;
   name: string;
