+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/shared/containers/ProcessStep.tsx
@@ -7,8 +7,8 @@ import { AltinnAppHeader } from 'altinn-shared/components';
 import { AltinnAppTheme } from 'altinn-shared/theme';
 import { IParty } from 'altinn-shared/types';
 import { returnUrlToMessagebox } from 'altinn-shared/utils';
+import { IValidations, ITextResource } from 'src/types';
 import { IRuntimeState, ProcessSteps } from '../../types';
-import { IValidations, ITextResource } from '../../types/global';
 import ErrorReport from '../../components/message/ErrorReport';
 import Header from '../../components/process-step/Header';
 import NavBar from '../../components/process-step/NavBar';
