--- src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/shared/resources/profile/index.d.ts
@@ -1,11 +1,0 @@
-import { IParty } from 'altinn-shared/types';
-
-export interface IProfile {
-  userId: number;
-  userName: string;
-  phoneNumber: string;
-  email: string;
-  partyId: number;
-  party: IParty;
-  userType: number;
-}
