+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/shared/resources/queue/dataTask/dataTaskQueueSagas.ts
@@ -1,9 +1,9 @@
 import { SagaIterator } from 'redux-saga';
 import { all, call, take } from 'redux-saga/effects';
+import { IAltinnWindow } from 'src/types';
 import FormDataActions from '../../../../features/form/data/formDataActions';
 import DataModelActions from '../../../../features/form/datamodel/formDatamodelActions';
 import FormLayoutActions from '../../../../features/form/layout/formLayoutActions';
-import { IAltinnWindow } from '../../../../types/global';
 import { FETCH_APPLICATION_METADATA_FULFILLED } from '../../applicationMetadata/actions/types';
 import { GET_INSTANCEDATA_FULFILLED } from '../../instanceData/get/getInstanceDataActionTypes';
 import QueueActions from '../queueActions';
