+++ src/development/LocalTest/appsettings.json
@@ -36,8 +36,8 @@
   "AllowedHosts": "*",
   "LocalPlatformSettings": {
     "LocalTestingStorageBasePath": "c:/AltinnPlatformLocal/",
-    "AppRepsitoryBasePath": "<Path to your repos folder where you have cloned the Altinn Studio application repositories>",
-    "LocalTestingStaticTestDataPath": "<Path to altinn-studio repo>/src/development/TestData/"
+    "AppRepsitoryBasePath": "c:/repos/",
+    "LocalTestingStaticTestDataPath": "c:/repos/altinn-studio/src/development/TestData/"
   },
   "PlatformSettings": {
     "ApiAuthorizationEndpoint": "http://localhost:5101/authorization/api/v1/"
